Patch 0667
・フリップザコインのコインタイマーに関する修正（pc.c, unit.c）
	-> ログイン時、コインタイマーが初期化されてなかったのを修正
	-> ログアウト時、コインタイマーが削除されてなかったのを修正
・ログイン時、AutoActiveタイマーを初期化するように修正（pc.c）
・bAutoActiveItem系において、効果時間が32768msまでしか指定できないので、
  activeitem構造体の tick メンバをshort型からint型に変更（map.h, pc.*）

+++ src/map/map.h
@@ -522,7 +522,7 @@ struct map_session_data {
 	struct {
 		int id[MAX_ACTIVEITEM];
 		short rate[MAX_ACTIVEITEM];
-		short tick[MAX_ACTIVEITEM];
+		int tick[MAX_ACTIVEITEM];
 		unsigned long flag[MAX_ACTIVEITEM];
 		short count;
 	} activeitem;
