+++ src/map/pc.h
@@ -97,7 +97,7 @@ int pc_dropitem(struct map_session_data*, int, int);
 
 int pc_checkweighticon(struct map_session_data *sd);
 
-int pc_activeitem(struct map_session_data* sd,int id,short rate,short tick,unsigned long flag);
+int pc_activeitem(struct map_session_data* sd,int id,short rate,int tick,unsigned long flag);
 int pc_activeitem_start(struct map_session_data* sd,unsigned long mode,unsigned int tick);
 
 int pc_bonus(struct map_session_data*,int,int);
